pyproject.toml
[tool.poetry]
name = "curlify2"
version = "1.0.0"
description = "Library to convert python requests and httpx object to curl command."
authors = ["Marcus Pereira <hi@marcuspereira.xyz>"]
license = "MIT"
readme = "README.md"

repository = "https://github.com/marcuxyz/curlify2"
homepage = "https://github.com/marcuxyz/curlify2"
keywords = ["python", "curl", "requests", "curlify", "httpx", "convert to curl"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.24.0"
responses = "^0.12.0"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
responses = "^0.12.0"
black = "^20.8b1"
pytest-httpx = "^0.10.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
